ARN-2292 Change versionNumber to an Int and fix tests

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsstrengthsbasedneedsassessmentsapi/oasys/controller/request/CounterSignAssessmentRequest.kt
@@ -8,7 +8,7 @@ import uk.gov.justice.digital.hmpps.hmppsstrengthsbasedneedsassessmentsapi.persi
 class CounterSignAssessmentRequest(
   @Schema(description = "The SAN Assessment version number that was returned from the Sign Assessment API call.", example = "2")
   @PositiveOrZero
-  val sanVersionNumber: Long,
+  val sanVersionNumber: Int,
   @Schema(description = "Indicates what type of case this is")
   val outcome: Tag,
   @Schema(description = "OASys User")
